refactor(types): Narrow readings to wire widths
The SDK converts none of the values it observes, so carrying them
as double added no precision. TactileState is uint16_t and
ActuatorState holds int32 position and velocity with int16
current, which takes HandState from 2448 bytes to 1032.

JointState::velocity_rad_s and effort_Nm go with them. Nothing in
the SDK ever wrote those, so a consumer could not tell a
placeholder zero from a joint at rest.

-Wall -Wextra comes in with the change, because it turns every
printf over these fields into a format mismatch and the build had
no warning flags to report one.

+++ cpp/examples/15_tactile.cpp
@@ -63,12 +63,12 @@ int main(int argc, char** argv)
     const HandState first = hand.get_state();
     std::printf("[example] raw values, %s fingertip taxels 0-5:", finger_name[1]);
     for (std::size_t t = 0; t < 6; ++t) {
-      std::printf(" %7.0f", first.tactile.fingers[1][t]);
+      std::printf(" %7d", first.tactile.fingers[1][t]);
     }
     std::printf("\n");
     std::printf("[example] raw values, palm taxels 0-5:        ");
     for (std::size_t t = 0; t < 6; ++t) {
-      std::printf(" %7.0f", first.tactile.palm[t]);
+      std::printf(" %7d", first.tactile.palm[t]);
     }
     std::printf("\n\n[example] none of those say whether anything is touching the hand\n\n");
 
@@ -83,11 +83,11 @@ int main(int argc, char** argv)
       const HandState state = hand.get_state();
       for (std::size_t f = 0; f < kFingerCount; ++f) {
         for (std::size_t t = 0; t < kTactileTaxelsPerFinger; ++t) {
-          finger_baseline[f][t] += state.tactile.fingers[f][t] / kSamples;
+          finger_baseline[f][t] += static_cast<double>(state.tactile.fingers[f][t]) / kSamples;
         }
       }
       for (std::size_t t = 0; t < kPalmTactileCount; ++t) {
-        palm_baseline[t] += state.tactile.palm[t] / kSamples;
+        palm_baseline[t] += static_cast<double>(state.tactile.palm[t]) / kSamples;
       }
       std::this_thread::sleep_for(std::chrono::milliseconds(50));
     }
